Fixed Project name typo(s), mantle works (buggy)
Mantling works. Delivered on time. Also, the project name was Infinte [sic] Skies for a long time, so I fixed it. I fixed the directory error in a previous commit and just continued with thsi commit.

Mantling code is a bit messy and un-commented but seems to work for now, which is a giant win. Animation doesn't super fit and is hilariously broken but I've been told not to worry about it.